requirements.txt
bpython
pypcapkit[Scapy]
pyshark
